requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tensorflow>=2.15.0
keras>=3.0.0
tokenizers>=0.15.0
nltk>=3.8.0
PyYAML>=6.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
imbalanced-learn>=0.11.0
pytest>=7.0.0
jupyter>=1.0.0
notebook>=6.5.0
wandb>=0.16.0